# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Cat Sitter API
  version: 1.0.0
  description: |
    API contract for Cat Sitter platform (sitter/client dual-role).
    - Auth: JWT bearer token, stateless backend (JJWT implementation).
    - Time format: UTC ISO-8601 (`date-time`, Instant mapping).
servers:
  - url: http://localhost:8080
    description: Local
  - url: https://api-uat.example.com
    description: UAT
  - url: https://api.example.com
    description: Production
...
paths:
...
  /api/v1/bookings/{bookingId}/reschedule:
    parameters:
      - $ref: "#/components/parameters/BookingId"
    patch:
      tags: [Booking]
      operationId: rescheduleBooking
      summary: Reschedule before quote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RescheduleBookingRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /api/v1/sitters/me/bookings:
    get:
      tags: [Booking]
      operationId: listSitterBookings
      summary: List sitter bookings
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/BookingStatus"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
...
  /api/v1/bookings/{bookingId}/cancel:
    parameters:
      - $ref: "#/components/parameters/BookingId"
    post:
      tags: [Booking]
      operationId: cancelBooking
      summary: Cancel booking
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelBookingRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /api/v1/bookings/{bookingId}/referrals:
    parameters:
      - $ref: "#/components/parameters/BookingId"
    post:
      tags: [TrustCircle]
      operationId: createReferral
      summary: Refer booking to trusted sitter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReferralRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Referral"
    get:
      tags: [TrustCircle]
      operationId: listReferrals
      summary: List referral records
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Referral"
  /api/v1/sitters/me/trust-circle/search:
    get:
      tags: [TrustCircle]
      operationId: searchTrustCircleCandidates
      summary: Search sitter candidates
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 100
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SitterSearchResult"
  /api/v1/sitters/me/trust-circle:
    post:
      tags: [TrustCircle]
      operationId: addTrustCirclePartner
      summary: Add trust-circle partner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddTrustCirclePartnerRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrustCircleRelation"
  /api/v1/sitters/me/trust-circle/{partnerId}:
    parameters:
      - $ref: "#/components/parameters/PartnerId"
    patch:
      tags: [TrustCircle]
      operationId: updateTrustCircleStatus
      summary: Update trust-circle relation status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTrustCircleStatusRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrustCircleRelation"
    delete:
      tags: [TrustCircle]
      operationId: removeTrustCirclePartner
      summary: Remove trust-circle partner
      responses:
        "204":
          description: No Content
...
  /api/v1/visits/{visitId}/checklist/custom-items:
    parameters:
      - $ref: "#/components/parameters/VisitId"
    post:
      tags: [Visit]
      operationId: createCustomChecklistItem
      summary: Add custom checklist item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomChecklistItemRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisitChecklistItem"
  /api/v1/visits/{visitId}/media:
    parameters:
      - $ref: "#/components/parameters/VisitId"
    post:
      tags: [Visit]
      operationId: uploadVisitMedia
      summary: Upload visit media metadata
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadVisitMediaRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisitMedia"
  /api/v1/visits/{visitId}/notes:
    parameters:
      - $ref: "#/components/parameters/VisitId"
    put:
      tags: [Visit]
      operationId: updateVisitNotes
      summary: Update sitter notes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVisitNotesRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visit"
...
  /api/v1/orders/{orderId}/complete:
    parameters:
      - $ref: "#/components/parameters/OrderId"
    post:
      tags: [Visit]
      operationId: completeOrder
      summary: Complete order after all visits done
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /api/v1/sitters/me/payouts:
    get:
      tags: [Billing]
      operationId: getMonthlyPayout
      summary: Monthly payout summary
      parameters:
        - name: month
          in: query
          required: true
          schema:
            type: string
            pattern: "^[0-9]{4}-[0-9]{2}$"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyPayout"
...
  /api/v1/sitters/me/subscription/checkout:
    post:
      tags: [Billing]
      operationId: createSubscriptionCheckout
      summary: Create checkout session for subscription plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubscriptionCheckoutRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSession"
...
  /api/v1/notifications:
    get:
      tags: [Notification]
      operationId: listNotifications
      summary: List notifications
      parameters:
        - name: unreadOnly
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /api/v1/notifications/{notificationId}/read:
    parameters:
      - $ref: "#/components/parameters/NotificationId"
    patch:
      tags: [Notification]
      operationId: markNotificationAsRead
      summary: Mark notification as read
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    BookingId:
      name: bookingId
      in: path
      required: true
      schema: { type: string, format: uuid }
    OrderId:
      name: orderId
      in: path
      required: true
      schema: { type: string, format: uuid }
    VisitId:
      name: visitId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    PartnerId:
      name: partnerId
      in: path
      required: true
      schema: { type: string, format: uuid }
    NotificationId:
      name: notificationId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
...
    BookingStatus:
      type: string
      enum: [PENDING, QUOTED, CONFIRMED, COMPLETED, CANCELLED]
    PaymentStatus:
      type: string
      enum: [UNPAID, PROOF_UPLOADED, PAID, REFUNDED]
    VisitStatus:
      type: string
      enum: [SCHEDULED, IN_PROGRESS, DONE, CANCELLED]
    TrustCircleStatus:
      type: string
      enum: [ACTIVE, BLOCKED]
...
    MediaType:
      type: string
      enum: [PHOTO, VIDEO]
...
    RescheduleBookingRequest:
      type: object
      required: [startAt, endAt]
      properties:
        startAt: { type: string, format: date-time }
        endAt: { type: string, format: date-time }
...
    CancelBookingRequest:
      type: object
      properties:
        reason: { type: string, maxLength: 500 }
    Booking:
      type: object
      required: [bookingId, orderStatus, paymentStatus, totalAmount]
      properties:
        bookingId: { type: string, format: uuid }
        orderStatus: { $ref: "#/components/schemas/BookingStatus" }
        paymentStatus: { $ref: "#/components/schemas/PaymentStatus" }
        baseAmount: { type: number }
        surchargeAmount: { type: number }
        discountAmount: { type: number }
        totalAmount: { type: number }
        pricingNotes: { type: string, nullable: true }
...
    CreateReferralRequest:
      type: object
      required: [toSitterId]
      properties:
        toSitterId: { type: string, format: uuid }
        reason: { type: string, maxLength: 500 }
    Referral:
      type: object
      required: [referralId, bookingId, fromSitterId, toSitterId, status]
      properties:
        referralId: { type: string, format: uuid }
        bookingId: { type: string, format: uuid }
        fromSitterId: { type: string, format: uuid }
        toSitterId: { type: string, format: uuid }
        status: { type: string, enum: [PENDING, ACCEPTED, REJECTED, EXPIRED] }
    SitterSearchResult:
      type: object
      required: [sitterId, name]
      properties:
        sitterId: { type: string, format: uuid }
        name: { type: string }
        avatarUrl: { type: string, format: uri, nullable: true }
    AddTrustCirclePartnerRequest:
      type: object
      required: [trustedSitterId]
      properties:
        trustedSitterId: { type: string, format: uuid }
    UpdateTrustCircleStatusRequest:
      type: object
      required: [status]
      properties:
        status: { $ref: "#/components/schemas/TrustCircleStatus" }
    TrustCircleRelation:
      type: object
      required: [relationId, ownerSitterId, trustedSitterId, status]
      properties:
        relationId: { type: string, format: uuid }
        ownerSitterId: { type: string, format: uuid }
        trustedSitterId: { type: string, format: uuid }
        status: { $ref: "#/components/schemas/TrustCircleStatus" }
    Visit:
      type: object
      required: [visitId, bookingId, status, visitStartTime, visitEndTime]
      properties:
        visitId: { type: string, format: uuid }
        bookingId: { type: string, format: uuid }
        status: { $ref: "#/components/schemas/VisitStatus" }
        visitStartTime: { type: string, format: date-time }
        visitEndTime: { type: string, format: date-time }
        sitterNotes: { type: string, nullable: true }
    VisitChecklistItem:
      type: object
      required: [itemId, itemName, isCustom, isCompleted]
      properties:
        itemId: { type: string, format: uuid }
        itemName: { type: string }
        isCustom: { type: boolean }
        isCompleted: { type: boolean }
        sortOrder: { type: integer }
...
    CreateCustomChecklistItemRequest:
      type: object
      required: [itemName]
      properties:
        itemName: { type: string, minLength: 1, maxLength: 255 }
        sortOrder: { type: integer, minimum: 0, default: 0 }
    UploadVisitMediaRequest:
      type: object
      required: [mediaType, mediaUrl]
      properties:
        mediaType: { $ref: "#/components/schemas/MediaType" }
        mediaUrl: { type: string, format: uri }
        caption: { type: string, maxLength: 500 }
    VisitMedia:
      type: object
      required: [visitMediaId, visitId, mediaType, mediaUrl, createdAt]
      properties:
        visitMediaId: { type: string, format: uuid }
        visitId: { type: string, format: uuid }
        mediaType: { $ref: "#/components/schemas/MediaType" }
        mediaUrl: { type: string, format: uri }
        caption: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    UpdateVisitNotesRequest:
      type: object
      required: [sitterNotes]
      properties:
        sitterNotes: { type: string, minLength: 1, maxLength: 4000 }
...
    MonthlyPayout:
      type: object
      required: [month, totalIncome, totalPaidOrders]
      properties:
        month: { type: string, pattern: "^[0-9]{4}-[0-9]{2}$" }
        totalIncome: { type: number }
        totalPaidOrders: { type: integer, minimum: 0 }
        records:
          type: array
          items:
            type: object
            required: [bookingId, amount]
            properties:
              bookingId: { type: string, format: uuid }
              amount: { type: number }
...
    CreateSubscriptionCheckoutRequest:
      type: object
      required: [planCode]
      properties:
        planCode: { type: string, minLength: 1, maxLength: 50 }
    CheckoutSession:
      type: object
      required: [checkoutUrl]
      properties:
        checkoutUrl: { type: string, format: uri }
    Notification:
      type: object
      required: [notificationId, type, title, isRead, createdAt]
      properties:
        notificationId: { type: string, format: uuid }
        type: { type: string }
        title: { type: string }
        body: { type: string, nullable: true }
        isRead: { type: boolean }
        readAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
...
security:
  - bearerAuth: []